back/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
google-cloud-aiplatform>=1.36.0
google-generativeai>=0.3.0
pydantic>=2.4.2
python-multipart>=0.0.6 